Move creation of HealthCheck objects to avoid concurrent map write errors. Make the events test non-parallel so that it does not affect other tests.
The Health Check tests cases write to a map within the health check
object. The test cases are parallel and they share a health check
instance, so they concurrently write to a single map without using
a mutex. By moving health check creation to within the test cases,
concurrent write is avoided.

The events test creates more SslCertificate resources that the quota
allows. It should not be executed in parallel with other tests that need
to create SslCertificates to test other features.

Change-Id: I878adfb7a228db7d4edb34f0fd867d3b2d0ff4c1